uso de bigint

+++ querys/stores procedures/Registros, ediciones y eliminaciones/asociado/sp_registrar_asociado.sql
@@ -12,7 +12,7 @@ AS
 BEGIN
     SET NOCOUNT ON;
 
-    DECLARE @nuevo_id INT;
+    DECLARE @nuevo_id BIGINT;
 
     BEGIN TRY
         BEGIN TRANSACTION;
